+++ app/Http/Livewire/MaterialIssuedJournal/MaterialIssuedJournalComponent.php
@@ -11,6 +11,10 @@ class MaterialIssuedJournalComponent extends Component
 {
     public function destroy($mid, $jid)
     {
+        if (!auth()->user()->can('material-journal-delete')) {
+            abort(404);
+        }
+
         $jev = JournalEntryVoucher::find($jid);
         $jev->delete();
 
@@ -23,6 +27,10 @@ class MaterialIssuedJournalComponent extends Component
 
     public function render()
     {
+        if (!auth()->user()->can('material-journal-show')) {
+            abort(404);
+        }
+
         $mijs = DB::table('material_issued_journals')
             ->join('journal_entry_vouchers', 'journal_entry_vouchers.code_id', '=', 'material_issued_journals.id')
             ->select('material_issued_journals.id as mid', 'rsmi_no', 'journal_entry_vouchers.jv_date as jdate', 'journal_entry_vouchers.jev_no as jno', 'journal_entry_vouchers.id as jid', 'journal_entry_vouchers.particulars as part')
